+++ test/exactly_lib_test/test_case_utils/string_transformers/case_converters.py
@@ -1,7 +1,8 @@
 import unittest
 
-from exactly_lib.test_case_utils.string_transformer.impl import case_converters as sut
-from exactly_lib_test.type_system.logic.string_model.test_resources import string_models
+from exactly_lib_test.test_case_utils.string_transformers.test_resources import argument_syntax as args
+from exactly_lib_test.test_case_utils.string_transformers.test_resources import case_converters as tr
+from exactly_lib_test.test_case_utils.string_transformers.test_resources.case_converters import CaseConverterConfig
 
 
 def suite() -> unittest.TestSuite:
@@ -11,101 +12,19 @@ def suite() -> unittest.TestSuite:
     ])
 
 
-class TestToUpper(unittest.TestCase):
-    def test_SHOULD_not_be_identity_transformer(self):
-        transformer = sut.ToUpperCaseStringTransformer('arbitrary custom')
-        self.assertFalse(transformer.is_identity_transformer)
-
-    def test_every_line_SHOULD_be_transformed(self):
-        # ARRANGE #
-        input_lines = [
-            'I object!',
-            'Object Oriented',
-            'Unidentified FLYING Object',
-        ]
-        model = string_models.of_lines(input_lines)
-        transformer = sut.ToUpperCaseStringTransformer('arbitrary custom')
-        # ACT #
-
-        actual = transformer.transform(model)
-
-        # ASSERT #
-
-        actual_lines = string_models.as_lines_list(actual)
-
-        expected_lines = [
-            'I OBJECT!',
-            'OBJECT ORIENTED',
-            'UNIDENTIFIED FLYING OBJECT',
-        ]
-
-        self.assertEqual(expected_lines,
-                         actual_lines)
-
-    def test_no_lines(self):
-        # ARRANGE #
-        input_lines = []
-        model = string_models.of_lines(input_lines)
-        transformer = sut.ToUpperCaseStringTransformer('arbitrary custom')
-        # ACT #
-
-        actual = transformer.transform(model)
-
-        # ASSERT #
-
-        actual_lines = string_models.as_lines_list(actual)
-
-        expected_lines = []
-
-        self.assertEqual(expected_lines,
-                         actual_lines)
+class TestToLower(tr.CaseConverterTestBase):
+    @property
+    def config(self) -> CaseConverterConfig:
+        return CaseConverterConfig(
+            args.to_lower_case(),
+            str.lower,
+        )
 
 
-class TestToLower(unittest.TestCase):
-    def test_SHOULD_not_be_identity_transformer(self):
-        transformer = sut.ToLowerCaseStringTransformer('arbitrary custom')
-        self.assertFalse(transformer.is_identity_transformer)
-
-    def test_every_line_SHOULD_be_transformed(self):
-        # ARRANGE #
-        input_lines = [
-            'I object!',
-            'Object Oriented',
-            'Unidentified FLYING Object',
-        ]
-        model = string_models.of_lines(input_lines)
-        transformer = sut.ToLowerCaseStringTransformer('arbitrary custom')
-        # ACT #
-
-        actual = transformer.transform(model)
-
-        # ASSERT #
-
-        actual_lines = string_models.as_lines_list(actual)
-
-        expected_lines = [
-            'i object!',
-            'object oriented',
-            'unidentified flying object',
-        ]
-
-        self.assertEqual(expected_lines,
-                         actual_lines)
-
-    def test_no_lines(self):
-        # ARRANGE #
-        input_lines = []
-        model = string_models.of_lines(input_lines)
-        transformer = sut.ToLowerCaseStringTransformer('arbitrary custom')
-        # ACT #
-
-        actual = transformer.transform(model)
-
-        # ASSERT #
-
-        actual_lines = string_models.as_lines_list(actual)
-
-        expected_lines = []
-
-        self.assertEqual(expected_lines,
-                         actual_lines)
+class TestToUpper(tr.CaseConverterTestBase):
+    @property
+    def config(self) -> CaseConverterConfig:
+        return CaseConverterConfig(
+            args.to_upper_case(),
+            str.upper,
+        )