+++ CHANGELOG.md
@@ -24,6 +24,7 @@ with exception of a "0." prefix until version 1 is released.
  - Exit codes - rename `IMPLEMENTATION_ERROR` -> `INTERNAL_ERROR`
  - Python compatibility: Increased to Python v 3.5.9
  - Instruction `conf`/`actor` - Remove "-" prefix from actor names
+ - String transformer - Case converters - Replaces builtin symbols with primitives
  
 ### Removed